refactor(shared): change the name of the service folder to services
