+++ src/shared/components/Drawer/components/Aside/Aside.tsx
@@ -3,7 +3,7 @@ import { useQuery } from "@tanstack/react-query";
 import React, { memo, useState } from "react";
 import useLocalStorage from "shared/common/hook/useLocalStorage";
 import AsideView from "./AsideView";
-import { fetchGetMenuService } from "shared/components/Drawer/components/Aside/service";
+import { fetchGetMenuService } from "shared/components/Drawer/components/Aside/services";
 import { ASIDE_QUERY_KEY } from "shared/components/Drawer/components/Aside/constants";
 
 const Aside: React.FC = () => {
