+++ src/shared/components/Drawer/components/Aside/components/LinkTreeItem/types/LinkTreeItem.types.ts
@@ -1,4 +1,4 @@
-import { IFetchGetMenuResponse } from "shared/components/Drawer/components/Aside/service/types";
+import { IFetchGetMenuResponse } from "shared/components/Drawer/components/Aside/services/types";
 
 type TypeSlug = "notes" | "texts";
