+++ src/shared/components/Drawer/components/Aside/services/fetchGetMenuService.ts
@@ -1,5 +1,5 @@
 import api from "shared/services/api";
-import { IFetchGetMenuResponse } from "shared/components/Drawer/components/Aside/service/types";
+import { IFetchGetMenuResponse } from "shared/components/Drawer/components/Aside/services/types";
 
 export const fetchGetMenuService = async () => {
 	const { data } = await api.get<IFetchGetMenuResponse>("/menu");
