+++ src/shared/components/Drawer/components/Aside/types/Aside.component.ts
@@ -1,5 +1,5 @@
 import { Palette } from "@mui/material";
-import { IFetchGetMenuResponse } from "shared/components/Drawer/components/Aside/service/types";
+import { IFetchGetMenuResponse } from "shared/components/Drawer/components/Aside/services/types";
 import { FunctionReturnsVoid } from "shared/common/types/AppTypes";
 
 export interface IAsideView {
